perf(util): remove allocations and improve speed
It's pretty interesting that there's a performance regression in "tens"
test.  I'm not entirely sure why, but I'll investigate it later.  Since
numbers in that range tend to be quite common, the slowdown isn't ideal.

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkRelNumber/hundreds_of_thousands-8     454           270           -40.50%
BenchmarkRelNumber/tens_of_thousands-8         507           275           -45.82%
BenchmarkRelNumber/thousands-8                 456           293           -35.66%
BenchmarkRelNumber/hundreds-8                  38.7          38.7          +0.05%
BenchmarkRelNumber/tens-8                      4.83          27.5          +468.18%

benchmark                                      old allocs     new allocs     delta
BenchmarkRelNumber/hundreds_of_thousands-8     3              0              -100.00%
BenchmarkRelNumber/tens_of_thousands-8         3              0              -100.00%
BenchmarkRelNumber/thousands-8                 3              0              -100.00%
BenchmarkRelNumber/hundreds-8                  1              0              -100.00%
BenchmarkRelNumber/tens-8                      0              0              +0.00%

benchmark                                      old bytes     new bytes     delta
BenchmarkRelNumber/hundreds_of_thousands-8     40            0             -100.00%
BenchmarkRelNumber/tens_of_thousands-8         40            0             -100.00%
BenchmarkRelNumber/thousands-8                 32            0             -100.00%
BenchmarkRelNumber/hundreds-8                  3             0             -100.00%
BenchmarkRelNumber/tens-8                      0             0             +0.00%
```

+++ modules/util/humanize_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 var relNumberCases = []struct {
 	name     string
-	input    int
+	input    int64
 	expected string
 }{
 	{"hundreds of thousands", 777777, "777.8k"},
